Downloads/resume parser concept/project/api/requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.4